+++ src/DS18B20_INT.cpp
@@ -28,7 +28,7 @@ DS18B20_INT::DS18B20_INT(OneWire* ow)
 {
   _oneWire      = ow;
   _addressFound = false;
-  _resolution   = TEMP_9_BIT;
+  _resolution   = 9;
 }
 
 
@@ -101,17 +101,11 @@ bool DS18B20_INT::getAddress(uint8_t* buf)
 }
 
 
-bool DS18B20_INT::setResolution(uint8_t bits)
+bool DS18B20_INT::setResolution(uint8_t resolution)
 {
   if (isConnected())
   {
-    switch (bits)
-    {
-      case 12: _resolution = TEMP_12_BIT;  break;
-      case 11: _resolution = TEMP_11_BIT;  break;
-      case 10: _resolution = TEMP_10_BIT;  break;
-      default: _resolution = TEMP_9_BIT;   break;
-    }
+    _resolution = resolution;
     _setResolution();
   }
   return _addressFound;
@@ -120,14 +114,7 @@ bool DS18B20_INT::setResolution(uint8_t bits)
 
 uint8_t DS18B20_INT::getResolution()
 {
-  switch (_resolution)
-  {
-    case TEMP_12_BIT: return 12;
-    case TEMP_11_BIT: return 11;
-    case TEMP_10_BIT: return 10;
-    case TEMP_9_BIT:  return 9;
-  }
-  return 0;
+  return _resolution;
 }
 
 
@@ -166,21 +153,26 @@ int16_t DS18B20_INT::_readRaw(void)
 }
 
 
-bool DS18B20_INT::_setResolution()
+void DS18B20_INT::_setResolution()
 {
-  if (_addressFound)
+  uint8_t res;
+  switch (_resolution)
   {
-    _oneWire->reset();
-    _oneWire->select(_deviceAddress);
-    _oneWire->write(WRITESCRATCH);
-    //  two dummy values for LOW & HIGH ALARM
-    _oneWire->write(0);
-    _oneWire->write(100);
+    case 12: res = TEMP_12_BIT;  break;
+    case 11: res = TEMP_11_BIT;  break;
+    case 10: res = TEMP_10_BIT;  break;
     //  lowest as default as we do only integer math.
-    _oneWire->write(_resolution);
-    _oneWire->reset();
+    default: res = TEMP_9_BIT;   break;
   }
-  return _addressFound;
+
+  _oneWire->reset();
+  _oneWire->select(_deviceAddress);
+  _oneWire->write(WRITESCRATCH);
+  //  two dummy values for LOW & HIGH ALARM
+  _oneWire->write(0);
+  _oneWire->write(100);
+  _oneWire->write(res);
+  _oneWire->reset();
 }